requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
boto3==1.26.95
botocore==1.29.95
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multidict==6.0.4
mypy==1.1.1
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pyee==9.0.4
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
